requirements.txt
streamlit
numpy
Pillow
tensorflow
torch
transformers @ git+https://github.com/huggingface/transformers.git
huggingface-hub
python-pptx
reportlab
